@@ -17,22 +17,22 @@
         }
     ],
     "require": {
-        "php": ">=7.3.0",
-        "symfony/framework-bundle": "^4.0|^5.0",
-        "thecodingmachine/safe": "^1.0",
-        "wernerdweight/ra": "^1.0|^2.0",
-        "wernerdweight/enhanced-exception": "^1.0",
-        "symfony/security-bundle": "^4.3|^5.0",
+        "php": ">=8.1",
+        "symfony/framework-bundle": "^6.0",
+        "thecodingmachine/safe": "^2.4",
+        "wernerdweight/ra": "^2.0",
+        "wernerdweight/enhanced-exception": "^2.0",
+        "symfony/security-bundle": "^6.0",
         "symfony/orm-pack": "^1.0|^2.0",
-        "wernerdweight/token-generator": "^1.0"
+        "wernerdweight/token-generator": "^2.0"
     },
     "require-dev": {
-        "wernerdweight/cs": "^2.0",
-        "thecodingmachine/phpstan-safe-rule": "^1.0",
-        "symfony/phpunit-bridge": "^5.0"
+        "wernerdweight/cs": "^3.0",
+        "thecodingmachine/phpstan-safe-rule": "^1.2",
+        "symfony/phpunit-bridge": "^6.0"
     },
     "suggest": {
-        "wernerdweight/doctrine-crud-api-bundle": "^1.0"
+        "wernerdweight/doctrine-crud-api-bundle": "^2.0"
     },
     "autoload": {
         "psr-4": {
@@ -49,13 +49,12 @@
         "phpstan": "phpstan analyse ./src/ ./tests/ --level max",
         "phpmd": "phpmd ./src/ text vendor/wernerdweight/cs/phpmd.xml",
         "ecs": "ecs check ./src/ ./tests/ --config ecs.php",
-        "ecs-fix": "ecs check ./src/ ./tests/ --config ecs.php --fix",
         "phpunit": "phpunit"
     },
     "extra": {
         "symfony": {
             "allow-contrib": false,
-            "require": "^4.3|^5.0"
+            "require": "^6.0"
         }
     }
 }
