@@ -5,9 +5,8 @@ namespace WernerDweight\ApiAuthBundle\Security;
 
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\EntityManagerInterface;
-use Doctrine\ORM\EntityRepository;
 use Symfony\Component\Security\Core\Exception\UnsupportedUserException;
-use Symfony\Component\Security\Core\Exception\UsernameNotFoundException;
+use Symfony\Component\Security\Core\Exception\UserNotFoundException;
 use Symfony\Component\Security\Core\User\UserInterface;
 use Symfony\Component\Security\Core\User\UserProviderInterface;
 use WernerDweight\ApiAuthBundle\Entity\ApiClientInterface;
@@ -16,16 +15,25 @@ use WernerDweight\ApiAuthBundle\Service\ConfigurationProvider;
 
 class ApiClientRefresher
 {
-    /** @var string */
+    /**
+     * @var string
+     */
     private const EXCEPTION_NOT_FOUND = 'There is no ApiClient for given client id!';
-    /** @var string */
+
+    /**
+     * @var string
+     */
     private const EXCEPTION_UNSUPPORTED_USER =
         '%s is not a supported authentication class. Make sure your class implements ApiClientInterface!';
 
-    /** @var EntityManager */
+    /**
+     * @var EntityManager
+     */
     private $entityManaager;
 
-    /** @var ConfigurationProvider */
+    /**
+     * @var ConfigurationProvider
+     */
     private $configurationProvider;
 
     /**
@@ -42,40 +50,31 @@ class ApiClientRefresher
     }
 
     /**
-     * @return EntityRepository<object>
-     */
-    private function getRepository(): EntityRepository
-    {
-        return $this->entityManaager->getRepository($this->configurationProvider->getClientClass());
-    }
-
-    /**
      * @throws \Safe\Exceptions\StringsException
      */
     public function refresh(UserInterface $user): ApiClientInterface
     {
-        if (!$user instanceof ApiClientInterface) {
+        if (! $user instanceof ApiClientInterface) {
             throw new UnsupportedUserException(\Safe\sprintf(self::EXCEPTION_UNSUPPORTED_USER, get_class($user)));
         }
 
-        $repository = $this->getRepository();
+        $repository = $this->entityManaager->getRepository($this->configurationProvider->getClientClass());
         if ($repository instanceof UserProviderInterface) {
             /** @var ApiClientInterface $apiClient */
             $apiClient = $repository->refreshUser($user);
             return $apiClient;
         }
 
-        $id = $this->entityManaager
-            ->getClassMetadata($this->configurationProvider->getClientClass())
-            ->getIdentifierValues($user);
-        if (true === empty($id)) {
+        $metadata = $this->entityManaager->getClassMetadata($this->configurationProvider->getClientClass());
+        $id = $metadata->getIdentifierValues($user);
+        if (0 === count($id)) {
             throw new ApiClientProviderException(ApiClientProviderException::EXCEPTION_NO_ID);
         }
 
         /** @var ApiClientInterface|null $apiClient */
         $apiClient = $repository->find($id);
         if (null === $apiClient) {
-            throw new UsernameNotFoundException(self::EXCEPTION_NOT_FOUND);
+            throw new UserNotFoundException(self::EXCEPTION_NOT_FOUND);
         }
         return $apiClient;
     }
